Haunted Places in the U.S.

A dataset that contains information about haunted locations in various cities across the United States.

Usage

haunted_data

Format

A data frame with several 10,992 rows and 11 columns, each describing a unique haunted location. The variables are as follows:

city

Name of the city where the haunted location is situated

location

Name or description of the haunted location

state

State abbreviation where the city is located

country

Country where the location is found (e.g., "United States")

description

Brief description of the haunting or history of the location

latitude

Latitude of the haunted location

longitude

Longitude of the haunted location

Source

https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2023/2023-10-10/haunted_places.csv'